requirements.txt
streamlit
pytesseract
pandas
numpy
Pillow
opencv-python
